refactor: add new functions and tests

+++ src/main/java/com/ronald/app/prueba/service/TareaServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class TareaServiceImpl implements ITareaService{
+public class TareaServiceImpl implements ITareaService {
 
     @Autowired
     private TareaRepository tareaRepository;
@@ -23,26 +23,31 @@ public class TareaServiceImpl implements ITareaService{
     @Override
     public Tarea buscarTareaPorId(Long id) {
         return tareaRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Entidad no encontrada con id: " + id));
+                .orElseThrow(() ->
+                        new EntityNotFoundException("Entidad no encontrada con id: " + id));
     }
 
     @Override
     public Tarea guardarTarea(Tarea tarea) {
-        if(tarea == null) throw  new NullPointerException("Entidad sin atributos");
+        if (tarea == null)
+            throw new NullPointerException("Entidad sin atributos");
+
         return tareaRepository.save(tarea);
     }
 
     @Override
     public void eliminarTarea(Long id) {
-        this.tareaRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Tarea no encontrada con id: "+ id));
+        if (!this.tareaRepository.existsById(id)) {
+            throw new EntityNotFoundException("Tarea no encontrada con id: " + id);
+        }
+        
         this.tareaRepository.deleteById(id);
     }
 
     @Override
-    public Long cantidadTareasByEstado(Estado estado ) {
-        if(estado==null ){
-            return    this.tareaRepository.count();
+    public Long cantidadTareasByEstado(Estado estado) {
+        if (estado == null) {
+            return this.tareaRepository.count();
         }
         return this.tareaRepository.countByEstado(estado);
     }
